[#180] Switched to StackBlurring
As of this commit, Quotograph will default to a blurring algorithm by
Mario Klingemann. This script does not rely on the RenderScript package,
works incredibly fast and with limited memory overhead. It appears to be
an ideal solution.

This also happens to bypass the StrictMode error caused by the
RenderScript packages.