[#172] Fixes Bad Line-Breaks
Turns out, we can't use StaticLayout.Builder yet because it's only
available in v23.
Back to the drawing board! We get the ending index of each word in the
quote and compare those indeces to the last visible character offset in
the StaticLayout. We shrink the font until each line ends on a known
word. Yah!